scripts/kb-audit/requirements.txt
openai>=1.30.0
faiss-cpu>=1.7.4
PyYAML>=6.0
python-dotenv>=1.0.0
pytest>=8.0.0
